% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{agedistribution}
\alias{agedistribution}
\title{Age distribution of individuals in Moria Camp (in bins)}
\format{
A vector with 12883 entries containing the age bin
of each individual
}
\source{
internal
}
\usage{
agedistribution
}
\description{
Age distribution in 10 year bins from individuals in Moria camp.
For each interval the lower range is inclusive and the upper range
is exclusive.
}
\keyword{datasets}
