% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_simulate.R
\name{calculate.max.edges}
\alias{calculate.max.edges}
\title{Quick fn to calculate max.possible amount of edges given some number
of nodes n}
\usage{
calculate.max.edges(n)
}
\arguments{
\item{n}{number of nodes}
}
\description{
Quick fn to calculate max.possible amount of edges given some number
of nodes n
}
