% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIQHRFNetModules.R
\name{custom.departures.net}
\alias{custom.departures.net}
\title{Departures: netsim Module}
\source{
EpiModel 2.0 source code
}
\usage{
custom.departures.net(dat, at)
}
\arguments{
\item{dat}{Master list object containing a \code{networkDynamic} object and other
initialization information passed from \code{\link{netsim}}.}

\item{at}{Current time step.}
}
\description{
This function simulates departure for use in \link{netsim} simulations.
}
\seealso{
\code{\link{netsim}}
}
\keyword{internal}
\keyword{netMod}
