% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiModelReplacement_OLD.R
\name{custom.get_prev.net.old}
\alias{custom.get_prev.net.old}
\title{Modification of the EpiModel::get_prev.net function}
\usage{
custom.get_prev.net.old(dat, at)
}
\description{
The performed modifications are clearly highlighted.
}
\details{
This function has been modified with to be able to store the different number
by vertex attribute (as per epi.by)

INTERNAL PURPOSE MOSTLY
}
\keyword{internal}
