% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiModelReplacement.R
\name{custom.init_status.net}
\alias{custom.init_status.net}
\title{Modification of the EpiModel::init_status.net function}
\source{
EpiModel 2.0
}
\usage{
custom.init_status.net(dat)
}
\description{
The performed modifications are clearly highlighted.
}
\details{
This function has been modified with the purpose of modifying how the times spent in each
state

INTERNAL PURPOSE MOSTLY
}
\keyword{internal}
