% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiModelReplacement_OLD.R
\name{custom.netsim.old}
\alias{custom.netsim.old}
\title{Modification of the EpiModel::netsim function}
\usage{
custom.netsim.old(x, param, init, control)
}
\description{
DEPRECATED (use custom.netsim instead)
The performed modifications are clearly highlighted.
}
\details{
INTERNAL PURPOSE MOSTLY
}
\keyword{internal}
