% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIQHRFNetModules.R
\name{custom.netsim}
\alias{custom.netsim}
\title{Modification of the EpiModel::netsim function}
\usage{
custom.netsim(x, param, init, control)
}
\description{
The performed modifications are clearly highlighted.
}
\details{
INTERNAL PURPOSE MOSTLY
}
\keyword{internal}
