% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiModelReplacement.R
\name{custom.process_out.net}
\alias{custom.process_out.net}
\title{Processes how output is saved}
\usage{
custom.process_out.net(dat_list)
}
\arguments{
\item{dat_list}{(From \code{EpiModel} docs) A list of Master objects in netsim simulations.}
}
\description{
This function was introduced in EpiModel 2 to simplify the code in netsim
}
