% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiModelReplacement_OLD.R
\name{custom.saveout.net.old}
\alias{custom.saveout.net.old}
\title{Modification of the EpiModel::saveout.net function}
\usage{
custom.saveout.net.old(dat, s, out = NULL)
}
\description{
The performed modifications are clearly highlighted.
}
\details{
This function has been modified with the purpose of modifying how the net object is saved
so that all of our custom states and their statistics can be saved.

INTERNAL PURPOSE MOSTLY
}
\keyword{internal}
