% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIQHRFNetModules.R
\name{exposure}
\alias{exposure}
\title{This function describes how people get EXPOSED
FROM THOSE INFECTED, QUARANTINED AND EXPOSED}
\usage{
exposure(dat, at)
}
\description{
This function describes how people get EXPOSED
FROM THOSE INFECTED, QUARANTINED AND EXPOSED
}
