% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIQHRFNetModules.R
\name{FromXToExposed}
\alias{FromXToExposed}
\title{1st Module to change:
Susceptible people get exposed instead of infected,
this script replaces infection.net}
\usage{
FromXToExposed(dat, at, finalProb, from.state = "i")
}
\description{
check ?infection.net for specification of what the base module does
I believe in current state this supports act.rate being vectors
}
