% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIQHRFNetModules.R
\name{infect}
\alias{infect}
\title{2nd module:
Infect: transition from exposed to infected, based on progress function from
EpiModel tutorial}
\usage{
infect(dat, at)
}
\description{
2nd module:
Infect: transition from exposed to infected, based on progress function from
EpiModel tutorial
}
