% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{net_example}
\alias{net_example}
\title{Example network object}
\format{
An object of class \code{list} of length 4.
}
\usage{
net_example
}
\description{
Object that is returned from running \code{net_simulate()}, available
for plot examples
}
\keyword{datasets}
