% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiModelReplacement.R
\name{netsim_loop}
\alias{netsim_loop}
\title{Internal function running the network simulation loop}
\source{
EpiModel2.0
}
\usage{
netsim_loop(x, param, init, control, s)
}
\description{
This function run the initialization and simulation loop for one
simulation. CNS note: this is not an exported object from the EpiModel package hence
I am pasting it here
}
\keyword{internal}
