% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{network.heatmap}
\alias{network.heatmap}
\title{Get network heatmap at a given time step}
\usage{
network.heatmap(net_build_object, at = 1)
}
\arguments{
\item{net_build_object}{objected returned by \code{net_simulate}}

\item{at}{timestep}
}
\description{
Get network heatmap at a given time step
}
