% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.states}
\alias{plot.states}
\title{Utility plotting functions
Plot epidemic curves for each state}
\usage{
\method{plot}{states}(net_build_object)
}
\description{
Utility plotting functions
Plot epidemic curves for each state
}
