% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_states}
\alias{plot_states}
\title{Plot epidemic curves for each state}
\usage{
plot_states(net_build_object)
}
\arguments{
\item{net_build_object}{object returned by \code{net_simulate}}
}
\description{
Plot epidemic curves for each state
}
\examples{
data(net_example)
plot_states(net_example)
}
