% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIQHRFNetModules.R
\name{quarantining}
\alias{quarantining}
\title{Transition bettween infected to quarantined state (one-to-one transition)}
\usage{
quarantining(dat, at)
}
\arguments{
\item{dat}{native Epimodel object}

\item{at}{simulation timestep}
}
\description{
Whether the transition occurs is governed by a Bernouilli trial with probability equal to ei.rate
}
\details{
This transition is useful to model population compliance with Public Health campaigns and their consequences.
}
