% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ratesbyage}
\alias{ratesbyage}
\title{Hospitalisation and Fatality rate for each age bin}
\format{
For each 10-year age bin a hospitalisation/fatality
rate is known (units: number of individuals/100/day)
\describe{
\item{AgeGroup}{bins of age with open lower end and closed upper end}
\item{hosp.rate}{Hospitalisation rate by age group (units: ?number of fatalities per case/1000)}
\item{fat.rate}{Fatality rate by age group (units: ?number of fatalities per case/1000)}
}
}
\usage{
ratesbyage
}
\description{
Hospitalisation and Fatality rate for each age bin
}
\keyword{datasets}
