% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIQHRFNetModules.R
\name{recover}
\alias{recover}
\title{Transition from quarantined, infected, requiring hospitalisation to recovered (multi-to-one transition)}
\usage{
recover(dat, at)
}
\arguments{
\item{dat}{native Epimodel object}

\item{at}{simulation timestep}
}
\description{
Whether the transition occurs is governed by a Bernouilli trial with probability equal to qr.rate or ir.rate or
hr.rate. Again this happens sequentially which results in same bias explained in the \strong{exposure} model
(see \code{help(exposure)})
}
