% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIQHRFNetModules.R
\name{RequireHospitalization}
\alias{RequireHospitalization}
\title{4th module:
I/Q to Require hospitalization}
\usage{
RequireHospitalization(dat, at)
}
\description{
4th module:
I/Q to Require hospitalization
}
