% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIQHRFNetModules.R
\name{custom.get_prev.net}
\alias{custom.get_prev.net}
\title{More source code: I've added this bit to be able to store the different number by vertex attribute
(as per epi.by)}
\usage{
custom.get_prev.net(dat, at)
}
\description{
More source code: I've added this bit to be able to store the different number by vertex attribute
(as per epi.by)
}
