% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIQHRFNetModules.R
\name{custom.init_status.net}
\alias{custom.init_status.net}
\title{Another one very similar to source code}
\usage{
custom.init_status.net(dat)
}
\description{
Another one very similar to source code
}
