% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIQHRFNetModules.R
\name{custom.netsim}
\alias{custom.netsim}
\title{For most this is netsim source code, I will make it clear where I have changed the code}
\usage{
custom.netsim(x, param, init, control)
}
\description{
For most this is netsim source code, I will make it clear where I have changed the code
}
