% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIQHRFNetModules.R
\name{custom_discord_edgelist}
\alias{custom_discord_edgelist}
\title{S, Susceptible
E, Exposed,
I, Infected
Q, Quarantined
H, Need hospitalisation
R, Recovered
F, Fatality}
\usage{
custom_discord_edgelist(dat, at, from = "i")
}
\description{
Network Modules
Based majorly in https://timchurches.github.io/blog/posts/2020-03-10-modelling-the-effects-of-public-health-interventions-on-covid-19-transmission-part-1/
work with ICM models
}
\details{
In this work we are going to redefine practically all base modules from EpiModel
}
