% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_simulate.R
\name{net_simulate}
\alias{net_simulate}
\title{Network simulation wrapper}
\usage{
net_simulate(
  n = 100,
  nsims = 3,
  nsteps = 90,
  prop.isobox = 0.43,
  iso.capacity = 10,
  tent.capacity = 4,
  verbose = F,
  cores = 1,
  d.rate = 0,
  external.contacts = 4,
  act.rate.se = 10,
  inf.prob.se = 0.02,
  act.rate.si = 10,
  inf.prob.si = 0.05,
  act.rate.sq = 2.5,
  inf.prob.sq = 0.05,
  ei.rate = 1/10,
  iq.rate = 1/30,
  ih.rate = 1/100,
  qh.rate = 1/100,
  hr.rate = 1/15,
  qr.rate = 1/20,
  ir.rate = 1/20,
  hf.rate = 1/50,
  hf.rate.overcap = 1/25,
  hosp.cap = 5,
  hosp.tcoeff = 0.5,
  a.rate = 0,
  di.rate = d.rate,
  ds.rate = d.rate,
  dr.rate = d.rate,
  i.num = 1,
  r.num = 0,
  e.num = 0,
  s.num = n - 1,
  f.num = 0,
  h.num = 0,
  q.num = 0,
  agedistribution = CampNetworkSimulator::agedistribution,
  ratesbyage = CampNetworkSimulator::ratesbyage
)
}
\arguments{
\item{n}{numbers of individuals, a random sample of this size will be taken from the Camp's population}

\item{nsims}{number of network simulations}

\item{nsteps}{number of timesteps (days, hours, weeks, arbitrary choice), the rates need to be defined in units
relative to this metric (e.g number of cases/day or /hour or /week)}

\item{prop.isobox}{Proportion of individuals in iso boxes, in this first implementation only two kinds of housing units
are considered (isoboxes and tent) -- can (and should) be changed in the future}

\item{iso.capacity}{Capacity of iso boxes}

\item{tent.capacity}{Capacity of tents}

\item{verbose}{whether function should be verbose in its command line output}

\item{cores}{number of cores to be used (default: number of available cores - 1)}

\item{d.rate}{departure rate (includes external deaths and actual departures) -- if set to 0 no invidual is consider to leave
or die from causes different than the epidemic agent}

\item{external.contacts}{Average number of external contacts that individuals make every day (external meaning from outside
their tent)}

\item{act.rate.se}{Act rate (contact rate) between susceptible and exposed individuals
-- default: 10 interactions of S-to-E/timestep/person}

\item{inf.prob.se}{infection probability of susceptible individuals upon contact with exposed individuals, these value
is often considered to be lower than that upon contact with infected individuals (though literature needs to be
checked for this) -- default: 0.02 (Out of every 100 contacts between S and E we expect 2 infections)}

\item{act.rate.si}{Act rate (contact rate) between susceptible and infected individuals
-- default: 10 interactions/timestep /person}

\item{inf.prob.si}{infection probability of susceptible individuals upon contact with susceptible individuals -- default: 0.05}

\item{act.rate.sq}{Act rate (contact rate) between susceptible and quarantined individuals
-- default: 2.5 interactions/timestep/person}

\item{inf.prob.sq}{infection probability of susceptible individuals upon contact with quarantined individuals -- default: 0.05}

\item{ei.rate}{Rate (can be seen as flow rate or probability) between exposed and infected individuals,
can take the value of a vector of length nsteps to simulate timevarying scenarios}

\item{iq.rate}{Rate between infected and quarantined states (~chances of quarantining when infected)}

\item{ih.rate}{Rate between infected and hospitalised states (~chances of requiring
hospitalisation when infected)}

\item{qh.rate}{Rate between quarantined and hospitalised states (~ chances of requiring hospitalisation
when quarantining)}

\item{hr.rate}{Rate between requiring hospitalisation and recovered states (~chances of recovering
when requiring hospitalisation)}

\item{qr.rate}{Rate between quarantined and recovered states (~chances of recovering when
in quarantined state)}

\item{ir.rate}{Rate between infected and recovered states (~chances of recovering when
in infected state)}

\item{hf.rate}{Rate between hospitalised and fatality state (~chances of dying from the infectious agent
given requiring hospitalisation)}

\item{hf.rate.overcap}{Same as hf.rate but when the hospital is at full capacity (As per hosp.cap)}

\item{hosp.cap}{Hospital capacity}

\item{hosp.tcoeff}{Increase in the hr.rate for every timestep spent in the hospital}

\item{a.rate}{arrival rate -> if set to 0, inflow of people is not considered (i.e. no births, no
immigration)}

\item{di.rate}{departure rate for infected individuals (defaults to d.rate)}

\item{ds.rate}{departure rate for susceptible individuals (defaults to d.rate)}

\item{dr.rate}{departure rate for susceptible individuals (defaults to d.rate)}

\item{i.num}{Initial number of infected individuals (defanetults to 1)}

\item{r.num}{Initial number of recovered individuals (defaults to 0)}

\item{e.num}{Initial number of exposed individuals (default to 0)}

\item{s.num}{Initial number of susceptible individuals (defaults to n-1)}

\item{f.num}{Initial number of fatalities (defaults to 0)}

\item{h.num}{Initial number of individuals requiring hospitalization (defaults to 0)}

\item{q.num}{Initial number of quarantined individuals (defaults to 0)}

\item{agedistribution}{vector containing the age distribution of
individuals in the camp. By default this will be loaded internally,
see \code{CampNetworkSimulator::agedistribution}}

\item{ratesbyage}{dataframe containing the hospitalisation rate and
fatality rate by age, by default \code{CampNetworkSimulator::ratesbyage} is used}
}
\description{
This function simulates a SEIQHRF epidemic on top of a
network of n individuals with a specific list of parameters as provided.
S - Susceptible
E - Exposed (also Infected)
I - Infected
Q - Quarantined (also Infected)
H - Requiring hospitalisation (also Infected)
R - Recovered
F - Fatality (due to infection)
}
\examples{

net_simulate(n= 100, nsims = 1, nsteps = 10)
}
