% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ratesbyage}
\alias{ratesbyage}
\title{Hospitalisation and Fatality rate for each age bin}
\format{
An object of class \code{data.frame} with 9 rows and 3 columns.
}
\usage{
ratesbyage
}
\description{
For each 10-year age bin a hospitalisation/fatality rate is known
(units: number of individuals/100/day)
}
\keyword{datasets}
