% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIQHRFNetModules.R
\name{RequireHospitalization}
\alias{RequireHospitalization}
\title{Transition between the infected or quarantined state to requiring hospitalisation (multi-to-one transition)}
\usage{
RequireHospitalization(dat, at)
}
\arguments{
\item{dat}{native Epimodel object}

\item{at}{simulation timestep}
}
\description{
This can be seen as analog to some sort of complication rate.
Whether the transition occurs is governed by a Bernouilli trial with probability equal to ih.rate or qh.rate
}
